Move Profilers class from index module to a separate one
